+++ backend/src/routes/index.ts
@@ -2,6 +2,7 @@
 import { Router } from "express";
 import { authRouter } from "../modules/auth/auth.routes";
 import { signupRouter } from "../modules/auth/signup.routes";
+import { loginRouter } from "../modules/auth/login.routes";
 
 const router = Router();
 
@@ -10,5 +11,6 @@ router.use("/signup", signupRouter);
 
 // Auth(Session)
 router.use("/auth", authRouter);
+router.use("/login", loginRouter);
 
 export default router;